Resolve new GCC printf-format warning in new log message text

+++ src/sbbs3/mailsrvr.c
@@ -3216,7 +3216,7 @@ static void smtp_thread(void* arg)
 			break;
 		if(strlen(buf) > 998) { /* RFC2822: "Each line of characters MUST be no more than 998 characters" */
 			lprintf(LOG_WARNING, "%04d %s %s sent an ILLEGALLY-LONG line (%d chars > 998): '%s'"
-				,socket, client.protocol, client_id, strlen(buf), buf);
+				,socket, client.protocol, client_id, (int)strlen(buf), buf);
 			sockprintf(socket, client.protocol, session, "500 Line too long");
 			break;
 		}
